Added mouse cursor
I used custom mouse cursor with sprites available in resources. Not sure about playback speed.
It is added in UI to be always visible.
It doesn't get along very well with camera movement on click.
Optional hide on right click lines are commented (they can be used to hide when hex menu is displayed).